+++ backend/Saay.Infrastructure/DTOs/TaskDTOs/AddTaskDto.cs
@@ -10,7 +10,7 @@ namespace Saay.Infrastructure.DTOs.TaskDTOs
 
         [Required]
         [Range(1, 3, ErrorMessage = "CategoryId must be between 1 and 3.")]
-        public int TaskCategoryId { get; set; }
+        public byte TaskCategoryId { get; set; }
 
         [Required]
         [StringLength(255, ErrorMessage = "Title must be at most 255 characters long.")]
